First version of a simple users management system.
It just lets us assign the users to a different mana-
geable role, as long as its current role is manageable
as well.
IMPORTANT: security will be changed in order to comply
with the requirement of custom permissions for each
role, so it has been completely disabled and everyone can
access to and do everything (sorry if I forgot to change
any parameter in any jsp).