License support improved.
Now we have some configurable values for the license management.
Two encoded licenses limit the number of terminals we can store
and the date limit. We also have a configurable key for decrypting
them, which is combined with a hardcoded one.
Also updated the configuration script with those new parameters.